+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 0.11.4
 
+* stax any - adds `--log` global flag to automatically execute `stax log` after finishing the requested command
 * stax pull - adds --stay-on-head flag to allow staying on the head/default branch after pulling
 
 0.11.3